Change task schema, fetch IP addresses and process
The tasks are now re-executed by themselves, avoiding un-syncing
problems within the async methos , which may cause problems when dealing
with a remote service.

Now the remote IP addresses of the TCP connections from the network are
collected and saved into the DB, so they are accesible for the future.

Code refactoring on the task and so on.